[packages/core][s]: nav and layout styleing v1
